refactor(bpu): move bpu modules into bpu package
- also fix BPU.scala import section
