+++ src/main/scala/xiangshan/frontend/bpu/FauFTB.scala
@@ -14,7 +14,7 @@
 * See the Mulan PSL v2 for more details.
 ***************************************************************************************/
 
-package xiangshan.frontend
+package xiangshan.frontend.bpu
 
 import chisel3._
 import chisel3.util._
@@ -22,6 +22,8 @@ import org.chipsalliance.cde.config.Parameters
 import scala.{Tuple2 => &}
 import utility._
 import xiangshan._
+import xiangshan.frontend.FullBranchPrediction
+import xiangshan.frontend.PrunedAddr
 
 trait FauFTBParams extends HasXSParameter with HasBPUConst {
   val numWays = 32
