+++ src/main/scala/xiangshan/frontend/bpu/ITTAGE.scala
@@ -23,18 +23,18 @@
  * Championship Branch Prediction (CBP). 2011.
  ***************************************************************************************/
 
-package xiangshan.frontend
+package xiangshan.frontend.bpu
 
 import chisel3._
 import chisel3.util._
 import freechips.rocketchip.util.SeqBoolBitwiseOps
 import org.chipsalliance.cde.config.Parameters
-import scala.{Tuple2 => &}
 import scala.math.min
 import utility._
 import utility.mbist.MbistPipeline
 import utility.sram.FoldedSRAMTemplate
 import xiangshan._
+import xiangshan.frontend._
 
 trait ITTageParams extends HasXSParameter with HasBPUParameter {
